backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pymongo==3.12.0
python-dotenv==0.19.0
python-multipart==0.0.5
motor==2.5.1
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.5
PyMuPDF==1.23.8
Pillow==10.2.0
python-jose==3.3.0
gunicorn==21.2.0